# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Invest Plus API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  fullName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Plan {
  id: number;
  name: string;
  slug: string;
  description: string;
  minAmount: number;
  /** @nullable */
  maxAmount: number | null;
  roiPercent: number;
  durationDays: number;
  features: string[];
  /** @nullable */
  badgeLabel: string | null;
}

export type InvestmentStatus = typeof InvestmentStatus[keyof typeof InvestmentStatus];


export const InvestmentStatus = {
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Investment {
  id: number;
  userId: number;
  planId: number;
  planName: string;
  amount: number;
  expectedReturn: number;
  status: InvestmentStatus;
  startDate: string;
  endDate: string;
  createdAt: string;
}

export interface InvestmentInput {
  planId: number;
  amount: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  return: 'return',
  fee: 'fee',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  /** @nullable */
  investmentId: number | null;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalInvested: number;
  totalReturns: number;
  activeInvestments: number;
  completedInvestments: number;
  portfolioValue: number;
  roiPercent: number;
  recentTransactions: Transaction[];
  activeInvestmentsList: Investment[];
}