# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Invest Plus API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: plans
    description: Investment plan operations
  - name: investments
    description: User investment operations
  - name: transactions
    description: Transaction history operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /plans:
    get:
      operationId: getPlans
      tags: [plans]
      summary: List all investment plans
      responses:
        "200":
          description: Investment plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"

  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a single investment plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /investments:
    get:
      operationId: getInvestments
      tags: [investments]
      summary: List current user investments
      responses:
        "200":
          description: User investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create a new investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Investment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get a specific investment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List current user transactions
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats for current user
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - fullName
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - email
        - password
        - fullName

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        roiPercent:
          type: number
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
        badgeLabel:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
        - description
        - minAmount
        - maxAmount
        - roiPercent
        - durationDays
        - features
        - badgeLabel

    Investment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        expectedReturn:
          type: number
        status:
          type: string
          enum: [active, completed, cancelled]
        startDate:
          type: string
        endDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - planId
        - planName
        - amount
        - expectedReturn
        - status
        - startDate
        - endDate
        - createdAt

    InvestmentInput:
      type: object
      properties:
        planId:
          type: integer
        amount:
          type: number
      required:
        - planId
        - amount

    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        investmentId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [deposit, withdrawal, return, fee]
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - investmentId
        - type
        - amount
        - description
        - createdAt

    DashboardSummary:
      type: object
      properties:
        totalInvested:
          type: number
        totalReturns:
          type: number
        activeInvestments:
          type: integer
        completedInvestments:
          type: integer
        portfolioValue:
          type: number
        roiPercent:
          type: number
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        activeInvestmentsList:
          type: array
          items:
            $ref: "#/components/schemas/Investment"
      required:
        - totalInvested
        - totalReturns
        - activeInvestments
        - completedInvestments
        - portfolioValue
        - roiPercent
        - recentTransactions
        - activeInvestmentsList