# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Invest Plus API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Investment } from './investment';
import type { Transaction } from './transaction';

export interface DashboardSummary {
  totalInvested: number;
  totalReturns: number;
  activeInvestments: number;
  completedInvestments: number;
  portfolioValue: number;
  roiPercent: number;
  recentTransactions: Transaction[];
  activeInvestmentsList: Investment[];
}