# lib/api-zod/src/generated/types/investment.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Invest Plus API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvestmentStatus } from './investmentStatus';

export interface Investment {
  id: number;
  userId: number;
  planId: number;
  planName: string;
  amount: number;
  expectedReturn: number;
  status: InvestmentStatus;
  startDate: string;
  endDate: string;
  createdAt: string;
}