# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Invest Plus API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  slug: string;
  description: string;
  minAmount: number;
  /** @nullable */
  maxAmount: number | null;
  roiPercent: number;
  durationDays: number;
  features: string[];
  /** @nullable */
  badgeLabel: string | null;
}